create kho model, fix type kinhdo, vido

+++ app/src/main/java/com/iot/logisticsapp/Model/CungCapHangHoa.java
@@ -15,8 +15,8 @@ public class CungCapHangHoa {
     public int tinhTrangVanChuyen;
 
     public String khoID;
-    public long kinhDo;
-    public long viDo;
+    public double kinhDo;
+    public double viDo;
 
 
     public CungCapHangHoa(String userID, String tenUser, String sdtUser, String diachiUser, String thoiGianDuKien, String tenItem, String tenLoaiItem, int khoiLuongItem, String diaChiCCC, String loaiHinhVanChuyen, int tinhTrangVanChuyen, String khoID, long kinhDo, long viDo) {
@@ -143,19 +143,19 @@ public class CungCapHangHoa {
         this.khoID = khoID;
     }
 
-    public long getKinhDo() {
+    public double getKinhDo() {
         return kinhDo;
     }
 
-    public void setKinhDo(long kinhDo) {
+    public void setKinhDo(double kinhDo) {
         this.kinhDo = kinhDo;
     }
 
-    public long getViDo() {
+    public double getViDo() {
         return viDo;
     }
 
-    public void setViDo(long viDo) {
+    public void setViDo(double viDo) {
         this.viDo = viDo;
     }
 }
